updates
Update to version vendored in irctest:

1. move scram.py to __init__.py, placating pyflakes
2. replace default nonce factory with `secrets`
